--- app/assets/stylesheets/start_stops.scss
@@ -1,3 +1,0 @@
-// Place all the styles related to the start_stops controller here.
-// They will automatically be included in application.css.
-// You can use Sass (SCSS) here: http://sass-lang.com/
