+++ app/assets/stylesheets/timecards.scss
@@ -1,3 +1,3 @@
-// Place all the styles related to the stops controller here.
+// Place all the styles related to the timecards controller here.
 // They will automatically be included in application.css.
 // You can use Sass (SCSS) here: http://sass-lang.com/
