--- app/controllers/start_stops_controller.rb
@@ -1,7 +1,0 @@
-class StartStopsController < ApplicationController
-
-  def index
-    @timecard = StartStop.all
-  end
-  
-end
