--- app/helpers/start_stops_helper.rb
@@ -1,2 +1,0 @@
-module StartStopsHelper
-end
