--- app/helpers/starts_helper.rb
@@ -1,2 +1,0 @@
-module StartsHelper
-end
