--- app/helpers/stops_helper.rb
@@ -1,2 +1,0 @@
-module StopsHelper
-end
