+++ app/helpers/timecards_helper.rb
@@ -1,0 +1,2 @@
+module TimecardsHelper
+end
