--- app/models/start.rb
@@ -1,3 +1,0 @@
-class Start < ApplicationRecord
-  belongs_to :stop
-end
