--- app/models/stop.rb
@@ -1,3 +1,0 @@
-class Stop < ApplicationRecord
-  belongs_to :start
-end
