+++ app/models/timecard.rb
@@ -1,0 +1,2 @@
+class Timecard < ApplicationRecord
+end
