--- app/views/stops/new.html.erb
@@ -1,1 +1,0 @@
-<p>投稿</p>
