+++ app/views/timecards/_start.html.erb
@@ -1,0 +1,3 @@
+<div class = buttons >
+  <%= button_to '出勤', 'new_start_stop_start_path', method: :get, class: "start-work" %>
+</div>
