+++ app/views/timecards/_stop.html.erb
@@ -1,0 +1,3 @@
+<div class = buttons >
+  <%= button_to '退勤', '#', method: :get, class: "start-work" %> 
+</div>
