+++ app/views/timecards/show.html.erb
@@ -1,0 +1,6 @@
+   <h1>時間管理</h1>
+          <% @start_stop.each do |start_stop| %>
+            <div class = clock >
+                 出勤<%= "#{start_stop.start_time}"%> 〜退勤<%= "#{start_stop.stop_time}" %>
+            </div>
+          <% end %>
