+++ config/routes.rb
@@ -1,10 +1,10 @@
 Rails.application.routes.draw do
   devise_for :users
   root to: "start_stops#index"
-  resources :start_stops do
-    resources :start, only: [:new, :create]
-    resources :stop, only: [:new, :create]
-  end
+  resources :start_stops, only: [:index, :show]
+  resources :start, only: [:new, :create]
+  resources :stop, only: [:new, :create]
+
