--- db/migrate/20220621124504_create_starts.rb
@@ -1,9 +1,0 @@
-class CreateStarts < ActiveRecord::Migration[6.0]
-  def change
-    create_table :starts do |t|
-      t.time :start_time, null: false
-
-      t.timestamps
-    end
-  end
-end
