--- db/migrate/20220627023137_create_stops.rb
@@ -1,9 +1,0 @@
-class CreateStops < ActiveRecord::Migration[6.0]
-  def change
-    create_table :stops do |t|
-      t.time :stop_time, null: false
-
-      t.timestamps
-    end
-  end
-end
