--- db/migrate/20220627024101_create_start_stops.rb
@@ -1,10 +1,0 @@
-class CreateStartStops < ActiveRecord::Migration[6.0]
-  def change
-    create_table :start_stops do |t|
-      t.references :user, null: false, foreign_key: true
-      t.references :start, null: false, foreign_key: true
-      t.references :stop, null: false, foreign_key: true
-      t.timestamps
-    end
-  end
-end
