+++ db/migrate/20220701074317_create_timecards.rb
@@ -1,0 +1,8 @@
+class CreateTimecards < ActiveRecord::Migration[6.0]
+  def change
+    create_table :timecards do |t|
+
+      t.timestamps
+    end
+  end
+end
