--- test/controllers/starts_controller_test.rb
@@ -1,7 +1,0 @@
-require 'test_helper'
-
-class StartsControllerTest < ActionDispatch::IntegrationTest
-  # test "the truth" do
-  #   assert true
-  # end
-end
