--- test/models/stop_test.rb
@@ -1,7 +1,0 @@
-require 'test_helper'
-
-class StopTest < ActiveSupport::TestCase
-  # test "the truth" do
-  #   assert true
-  # end
-end
